internal reworks

+++ rus_opus_preprocess.py
@@ -1,44 +1,51 @@
+import datetime
+import time
+
 import ffmpeg
 import os
 import argparse
 from multiprocessing import Pool, freeze_support
-from pathlib import Path
 from functools import partial
 
+
 parser = argparse.ArgumentParser(description='Dataset args')
-parser.add_argument("-d", "--dataset_path", type=Path, help="Path to your root dataset folder")
+parser.add_argument("-d", "--dataset_path", type=str, help="Path to your root dataset folder")
 
-#this file probably sucks but it worked when I used it, idk about now :D
+test_run = False
 
-def ffmpeg_convert(filei, dspath, dir1, dir2):
-    file, i = filei
+def ffmpeg_convert(dspath, dir1, dir2, filei):
+    i, file = filei
     fullname = dspath + dir1 + "\\" + dir2 + "\\" + file
     if fullname.split(".")[1] == "opus":
         cleandirname = str(i)
         while len(cleandirname) < 3:
             cleandirname = "0" + cleandirname
         wavname = dspath + dir1 + "\\" + dir2 + "\\" + "utterance-" + cleandirname + ".wav"
-        # print()
-        # print(fullname, "to", wavname)
         txtname = fullname.replace("opus", "txt")
         txtoutname = wavname.replace("wav", "txt")
-        # print(txtname, txtoutname)
-        # print()
-        (
-            ffmpeg  # the ffmpeg part, works like a charm but very slowly
-                .input(fullname)
-                .output(wavname)
-                .run()
-        )
-        # print(fullname)
-        os.remove(fullname)  # be careful
-        os.rename(txtname, txtoutname)
+        if test_run:
+            print(fullname, wavname)
+            print(txtname, txtoutname)
+        else:
+            (
+                ffmpeg  # the ffmpeg part, works like a charm but very slowly
+                    .input(fullname)
+                    .output(wavname, loglevel="quiet")
+                    .run()
+            )
+            os.remove(fullname)  # be careful
+            os.rename(txtname, txtoutname)
 
 def convert_opuses(dspath):
+    times = []
     for dir1 in os.listdir(dspath):
+        tim = time.time()
         for dir2 in os.listdir(dspath+dir1):
             pool = Pool()
             pool.map(partial(ffmpeg_convert, dspath, dir1, dir2), enumerate(os.listdir(dspath+dir1+"\\"+dir2)))
+        times.append(time.time() - tim)
+        print("time per dir {}: ".format(str(dir1)), datetime.timedelta(seconds=round(time.time() - tim, 2)))
+    print("overall time: ", datetime.timedelta(seconds=round(sum(times), 2)))
 
 def rename_secondary_dir(insecdir):
     j=0
@@ -49,8 +56,10 @@ def rename_secondary_dir(insecdir):
             s = "0"+s
         newpath = insecdir+"\\"+"book-"+s
         j+=1
-        os.rename(inbefore, newpath)
-        # print("\t"+inbefore, "to", newpath)
+        if test_run:
+            print(inbefore, newpath)
+        else:
+            os.rename(inbefore, newpath)
 
 def restore_structure(rpath):
     i=0
@@ -60,11 +69,21 @@ def restore_structure(rpath):
             s = "0"+s
         newpath = "speaker-"+s
         i+=1
-        os.rename(rpath+dirq, rpath+newpath)
-        rename_secondary_dir(rpath + "\\" + newpath)
+        if test_run:
+            print(rpath+dirq, rpath+newpath)
+        else:
+            os.rename(rpath+dirq, rpath+newpath)
+        rename_secondary_dir(rpath + newpath)
 
 if __name__=="__main__":
     freeze_support()
     args = parser.parse_args()
-    restore_structure(args.dataset_path)
-    convert_opuses(args.dataset_path)
+    pat = args.dataset_path if args.dataset_path[-1] == "\\" or "/" else args.dataset_path+"\\"
+    print("Working with path: ", pat)
+    if "speaker-000" in os.listdir(pat):
+        print("structure already restored previously, skipping")
+    else:
+        print("restoring structure..")
+        restore_structure(pat)
+    print("converting opuses..")
+    convert_opuses(pat)
